# (Breaking change) Faster Handle Scribe