--- a/katip/examples/example_lens.hs
+++ b/katip/examples/example_lens.hs
@@ -50,7 +50,7 @@ main = do
   -- messages entirely. Note that katip provides facilities like
   -- 'unregisterScribe' and 'registerScribe' to make it possible to
   -- hot-swap scribes at runtime if you need to.
-  handleScribe <- mkHandleScribe ColorIfTerminal stdout InfoS V2
+  (handleScribe, finaliser) <- mkHandleScribe ColorIfTerminal stdout InfoS V2
   let le' = registerScribe "stdout" handleScribe le
   let s = MyState M.mempty mempty le'
   runStack s $ do
@@ -68,6 +68,7 @@ main = do
     $(logTM) InfoS "Namespace and context are back to normal"
     noLogging $
       $(logTM) DebugS "You'll never see this log message!"
+  finaliser
 
 
 -------------------------------------------------------------------------------