--- a/katip/katip.cabal
+++ b/katip/katip.cabal
@@ -55,6 +55,7 @@ library
 
   build-depends: base >=4.5 && <5
                , aeson >=0.6 && <1.2
+               , async < 3.0.0.0
                , auto-update >= 0.1 && < 0.2
                , bytestring >= 0.9 && < 0.11
                , containers >=0.4 && <0.6
@@ -76,6 +77,7 @@ library
                , microlens >= 0.2.0.0 && < 0.5
                , microlens-th >= 0.1.0.0 && < 0.5
                , semigroups
+               , unagi-chan < 0.5.0.0
 
   hs-source-dirs:      src
   default-language:    Haskell2010