--- a/stack-lts-4.yaml
+++ b/stack-lts-4.yaml
@@ -11,6 +11,7 @@ extra-deps:
   - tagged-0.8.3
   - uuid-1.3.12
   - rng-utils-0.2.1
+  - unagi-chan-0.4.0.0
 flags:
   katip:
     lib-Werror: true