Hey @MichaelXavier ,

as per #46 I have played around with `unagi-chan` in order to speed up the performances of the `Handle` `Scribe`. Before delving into the details these are the results, where topmost is the old bench, bottom is the new implementation:

```
katip-0.3.1.4: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
benchmarking Katip.Scribes.Handle/Bytestring Builder
time                 5.759 μs   (5.683 μs .. 5.857 μs)
                     0.998 R²   (0.998 R² .. 0.999 R²)
mean                 5.755 μs   (5.703 μs .. 5.827 μs)
std dev              201.7 ns   (166.8 ns .. 233.3 ns)
variance introduced by outliers: 44% (moderately inflated)

Benchmark bench: FINISH

katip-0.3.1.4: benchmarks
Running 1 benchmarks...
Benchmark bench: RUNNING...
benchmarking Katip.Scribes.Handle/Bytestring Builder
time                 38.29 ns   (37.75 ns .. 39.00 ns)
                     0.998 R²   (0.996 R² .. 0.999 R²)
mean                 38.30 ns   (37.92 ns .. 38.87 ns)
std dev              1.442 ns   (1.000 ns .. 1.927 ns)
variance introduced by outliers: 59% (severely inflated)
```

As you can see, quite an improvement!

There are a couple of things to consider here though:

* The implementation is built for speed and right now it will drop messages if the queue is full (cfr `tryReadChan`)
* There is **absolutely no** story for resource cleanup, which means that if one stops abruptly the program and there are messages left in the queue, the async will die and those messages won't be delivered. I have already stubbed out a cleanup mechanism via a `PoisonPill` that we should deliver to the `worker` when we stop, but currently there is no way of doing that without either breaking the `mkHandleScribe` API or rework the internals of Katip (as per #46).

Therefore, I have issued this PR anyway but I would consider more a preview more than anything else, unless you think that:

1. The performance gains is such that we should merge this anyway
2. We could create a `mkHandleScribeUnsafe` logger which would use this implementation but warn the user about the implications
3. We use this as a start for a discussion but leave it unmerged.

Whatever you decide I'm cool with that!

A.