@MichaelXavier 2017-02-17
This looks really good, thank you! Its a shame unagi-chan doesn't have a closeable variant. Your `PoisonPill` variation is clever. I can't think of another way to do it because unagi-chan splits into 2 channels, an in and an out side. We could possibly create a wrapper like `data Closeable a = Closeable (MVar Bool) a`, share the mvar (or tvar or whatever is appropriate) between both sides such that write operations are a noop when an InChan is closed and a closed OutChan means its time to die. However, this is sounding a whole lot more complicated and probably worse performance than the semantics of a poison pill.

Handle skirted the issue of cleanup so far by taking a handle from input, thus shirking the responsibility of cleaning it up. Katip elastic search already dealt with this issue by [returning a finalizer](https://github.com/Soostone/katip/blob/master/katip-elasticsearch/src/Katip/Scribes/ElasticSearch.hs#L264). Perhaps we should bite the bullet and make that small API change now so that your handle scribe returns an action that enqueues the poison pill? I think somewhere in the docs I may even have guidelines for writing scribes so that would be worth adding a note that scribes should return a tuple with a finalizer if applicable.