@MichaelXavier 2017-02-17
One small thing worth mentioning is that katip-elasticsearch's finalizing function is blocking which is probably something we want to preserve. The idea is that you would put it at the end of your program's lifecycle and when it returns, that means all writes have been committed and its safe to close. I think you could do this simply by having your finalizer be something like `writeChan inChan PoisionPill >> void (waitCatchAsync worker)` or something like that.