@adinapoli 2017-02-17
Hey Michael!

> Perhaps we should bite the bullet and make that small API change now so that your handle scribe returns an action that enqueues the poison pill?

Believe it or not that was the solution I was leaning towards, but I refrained doing that as I thought breaking the API would have made you not so happy 😉 

Jokes aside, I think that is a very sensible idea and I can actually see a pattern emerging here (maybe something we could address as part of #49 ? ) It we would extend the `Scribe` type with not only the "push" action **and** a new field storing a finaliser of some sort, we could lift the finalisation story into the guts of Katip and relief the users from doing that. But maybe that's not that simple and we need to go back to the drawing board 😝 

As far as this PR is concerned, I'm very happy with whatever you decide: possibly the morally right thing to do would be here to bite the bullet and change the API of `mkHandleScribe` to return (Scribe, IO ()), so that at least we give the user the chance to have the piece of mind of having all the logs delivered?