@adinapoli 2017-02-20
I have been thinking about this, and I'm definitely going (if you are OK with that) to break the API and modify the `mkHandleScribe` signature to be `IO (Scribe, IO ())`.

One compelling reason I have found is that otherwise we have no way to make tests reproducible, and that's the reason of part of the travis failure (the other part is me being lazy and not adding the necessary deps to the stack-XX.yml files).

@MichaelXavier what do you think? 👍  or 👎  ?