@MichaelXavier 2017-02-20
This look great to me. I'd say keep the finalizer pair like you have it. My plan with this is to merge it into master and then hopefully solve 2 problems before a release:

1. Figure out a reasonably compatible strategy to get katip-elasticsearch on bloodhound 0.13 (it changed module structure)
2. Get this internal channel in katip built and have it delegate finalization/removal actions.

I hope you're ok with submoduling or otherwise vendoring katip in the meantime to get the improved performance. I've got a couple of production gremlins to troubleshoot before I can get to this other implementation work. It looks like the examples dir needs a quick update to account for the finalizer and then a few of the LTSes may need extra-deps in their stack.yaml files, but otherwise this looks sound to me! Thank you!